requirements.txt
streamlit
fastapi
uvicorn
pymupdf
pandas
python-dotenv
requests
python-docx
